Initial Subscription and multiversioning support
