+++ ddf-rpc-api/src/main/java/com/connexta/ddf/attribute/enumerations/EnumerationMethods.java
