+++ ddf-rpc-api/src/main/java/com/connexta/ddf/catalog/direct/ExtendedMethods.java
