+++ ddf-rpc-api/src/main/java/com/connexta/jsonrpc/Error.java
