+++ ddf-rpc-api/src/main/java/com/connexta/jsonrpc/Metadata.java
