+++ ddf-rpc-api/src/main/java/com/connexta/jsonrpc/impl/ErrorImpl.java
