+++ ddf-rpc-api/src/main/java/com/connexta/util/MapFactory.java
