+++ ddf-rpc-api/src/main/java/com/connexta/util/StringUtils.java
