+++ ddf-rpc-api/src/main/resources/OSGI-INF/metatype/metatype.xml
