+++ ddf-rpc-api/src/test/java/com/connexta/ddf/catalog/direct/ExtendedMethodsTest.java
