+++ email/2.19-email-impl/src/main/java/com/connexta/jsonrpc/email/impl/v_2_19/EmailImpl.java
@@ -1,0 +1,12 @@
+package com.connexta.jsonrpc.email.impl.v_2_19;
+
+import com.connexta.jsonrpc.email.EmailResolver;
+import ddf.security.SubjectUtils;
+import org.apache.shiro.SecurityUtils;
+
+public class EmailImpl implements EmailResolver {
+  @Override
+  public String getCurrentSubjectEmail() {
+    return SubjectUtils.getEmailAddress(SecurityUtils.getSubject());
+  }
+}
