+++ email/email-service/src/main/java/com/connexta/jsonrpc/email/EmailResolver.java
@@ -1,0 +1,6 @@
+package com.connexta.jsonrpc.email;
+
+public interface EmailResolver {
+
+  String getCurrentSubjectEmail();
+}
