+++ email/pom.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <parent>
+        <groupId>com.connexta</groupId>
+        <artifactId>ddf-jsonrpc</artifactId>
+        <version>0.5-SNAPSHOT</version>
+    </parent>
+
+    <groupId>com.connexta.jsonrpc.email</groupId>
+    <artifactId>email</artifactId>
+    <packaging>pom</packaging>
+    <name>Connexta :: JSON RPC :: Email</name>
+
+    <modules>
+        <module>email-service</module>
+        <module>2.26-email-impl</module>
+        <module>2.19-email-impl</module>
+    </modules>
+
+
+</project>
