+++ jsonrpc-features/src/main/resources/features.xml
@@ -1,0 +1,34 @@
+<!--
+/**
+ * Copyright (c) Connexta
+ *
+ * This is free software: you can redistribute it and/or modify it under the terms of the GNU Lesser General Public License as published by the Free Software Foundation, either
+ * version 3 of the License, or any later version.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU Lesser General Public License for more details. A copy of the GNU Lesser General Public License is distributed along with this program and can be found at
+ * <http://www.gnu.org/licenses/lgpl.html>.
+ *
+ **/
+ -->
+
+<features name="${project.artifactId}-${project.version}"
+  xmlns="http://karaf.apache.org/xmlns/features/v1.3.0"
+  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+  xsi:schemaLocation="http://karaf.apache.org/xmlns/features/v1.3.0 http://karaf.apache.org/xmlns/features/v1.3.0">
+
+  <feature name="jsonrpc-2.19" version="${project.version}"
+    description="JSONRPC for ddf version 2.19.x">
+    <bundle>mvn:com.connexta.jsonrpc.email/email-service/${project.version}</bundle>
+    <bundle>mvn:com.connexta.jsonrpc.email/2.19-email-impl/${project.version}</bundle>
+    <bundle>mvn:com.connexta/ddf-rpc-api/${project.version}</bundle>
+  </feature>
+
+  <feature name="jsonrpc-2.26" version="${project.version}"
+    description="JSONRPC for ddf version 2.19.x">
+    <bundle>mvn:com.connexta.jsonrpc.email/email-service/${project.version}</bundle>
+    <bundle>mvn:com.connexta.jsonrpc.email/2.26-email-impl/${project.version}</bundle>
+    <bundle>mvn:com.connexta/ddf-rpc-api/${project.version}</bundle>
+  </feature>
+
+</features>
