--- src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -1,123 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-    /**
-    * Copyright (c) Codice Foundation
-    *
-    * This is free software: you can redistribute it and/or modify it under the terms of the GNU Lesser General Public License as published by the Free Software Foundation, either
-    * version 3 of the License, or any later version.
-    *
-    * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
-    * See the GNU Lesser General Public License for more details. A copy of the GNU Lesser General Public License is distributed along with this program and can be found at
-    * <http://www.gnu.org/licenses/lgpl.html>.
-    *
-    **/
--->
-<blueprint xmlns="http://www.osgi.org/xmlns/blueprint/v1.0.0" xmlns:cm="http://aries.apache.org/blueprint/xmlns/blueprint-cm/v1.1.0">
-  
-  <reference id="catalogFramework" interface="ddf.catalog.CatalogFramework"/>
-  
-  <reference id="attributeRegistry" interface="ddf.catalog.data.AttributeRegistry"/>
-  
-  <reference-list id="metacardTypes" interface="ddf.catalog.data.MetacardType"/>
-  
-  <reference id="filterBuilder" interface="ddf.catalog.filter.FilterBuilder"/>
-  
-  <reference id="persistentStore" interface="org.codice.ddf.persistence.PersistentStore"/>
-  
-  <reference-list id="metacardActionProviders" interface="ddf.action.ActionProvider" filter="(id=catalog.data.metacard.*)"/>
-  
-  <reference-list id="metacardMultiActionProviders" interface="ddf.action.MultiActionProvider" filter="(id=catalog.data.metacard.*)"/>
-  
-  <bean id="metacardActionRegistry" class="ddf.action.impl.ActionRegistryImpl">
-    <argument ref="metacardActionProviders"/>
-    <argument ref="metacardMultiActionProviders"/>
-  </bean>
-  
-  <bean id="subscriptionMethods" class="com.connexta.ddf.persistence.subscriptions.SubscriptionMethods">
-    <argument ref="persistentStore"/>
-  </bean>
-  
-  <reference id="defaultAttributeValidatorRegistry" interface="ddf.catalog.validation.AttributeValidatorRegistry" availability="optional"/>
-  
-  <reference-list id="attributeInjectors" interface="ddf.catalog.data.AttributeInjector" availability="optional"/>
-
-  <reference id="inputTransformer" interface="ddf.catalog.transform.InputTransformer"
-  filter="(id=xml)"/>
-  
-  <bean id="enumerationExtractor" class="com.connexta.ddf.attribute.enumerations.EnumerationExtractor">
-    <argument ref="defaultAttributeValidatorRegistry"/>
-    <argument ref="metacardTypes"/>
-    <argument ref="attributeInjectors"/>
-  </bean>
-
-  <bean id="listHandler" class="com.connexta.ddf.transformer.RpcListHandler">
-    <argument ref="inputTransformer" />
-    <argument ref="catalogFramework"/>
-    <argument ref="metacardTypes" />
-  </bean>
-
-  <bean id="metacardMap" class="com.connexta.ddf.catalog.direct.MetacardMap">
-    <argument ref="listHandler" />
-  </bean>
-
-  <reference id="subjectOperations" interface="ddf.security.SubjectOperations" />
-
-  <bean id="catalogMethods" class="com.connexta.ddf.catalog.direct.CatalogMethods">
-    <argument ref="catalogFramework"/>
-    <argument ref="attributeRegistry"/>
-    <argument ref="metacardTypes"/>
-    <argument ref="filterBuilder"/>
-    <argument ref="metacardActionRegistry"/>
-    <argument ref="subscriptionMethods"/>
-    <argument ref="metacardMap" />
-    <argument ref="listHandler" />
-    <argument ref="subjectOperations" />
-    <!--    <cm:managed-properties persistent-id="com.connexta.ddf.resourcemanagement.usage"-->
-    <!--      update-strategy="container-managed"/>-->
-    <!--    <property name="monitorLocalSources" value="false"/>-->
-  </bean>
-
-  <bean id="extendedMethods" class="com.connexta.ddf.catalog.direct.ExtendedMethods">
-    <argument ref="filterBuilder"/>
-    <argument ref="catalogFramework"/>
-    <cm:managed-properties persistent-id="com.connexta.ddf.catalog.direct.ExtendedMethods" update-strategy="container-managed"/>
-    <property name="cloneableAssociationAttributes">
-      <list>
-        <value>queries</value>
-      </list>
-    </property>
-  </bean>
-
-  <bean id="enumerationMethods" class="com.connexta.ddf.attribute.enumerations.EnumerationMethods">
-    <argument ref="enumerationExtractor"/>
-    <property name="metacardTypes">
-      <reference-list interface="ddf.catalog.data.MetacardType" availability="optional"/>
-    </property>
-  </bean>
-
-  <bean id="jsonRpc" class="com.connexta.jsonrpc.impl.JsonRpc" >
-
-    <argument>
-      <list>
-        <ref component-id="catalogMethods"/>
-        <ref component-id="extendedMethods"/>
-        <ref component-id="enumerationMethods"/>
-      </list>
-    </argument>
-
-  </bean>
-
-  <reference-list interface="com.connexta.jsonrpc.MethodSet" availability="optional">
-    <reference-listener ref="jsonRpc" bind-method="bindMethods" unbind-method="unbindMethods"/>
-  </reference-list>
-
-  <bean id="jsonRpcHttpServlet" class="com.connexta.jsonrpc.impl.JsonRpcHttpServlet">
-    <argument ref="jsonRpc"/>
-  </bean>
-
-  <service ref="jsonRpcHttpServlet" interface="javax.servlet.Servlet">
-    <service-properties>
-      <entry key="urlPatterns" value="/direct/*"/>
-    </service-properties>
-  </service>
-</blueprint>
